Use the editingFinished event rather than focusChanged(false)
This change allows the user to stay on an edit control and press ENTER, avoiding
the potentially confusing situation where the user makes a change, stays on
the control and expects there to be an update (but the update didn't happen
until the user left the control). With the change, the update happens either
when the user presses ENTER or leaves the control, provided the control is
not empty.